Lisätty tuotteiden määrä ostoskoriin, jonka näkee headerissa, ei pävity dynaamisesti. Lisätty myös virheen tarkistusta, jotta asiakas ei voi maksaa 0 tuotetta

+++ Verkkokauppa/frontend/src/ShoppingCart.jsx
@@ -6,7 +6,8 @@ export default function ShoppingCart() {
     const [shoppingCart, setShoppingCart] = useState([]);
     const [error, setError] = useState("");
     const [notificationMessage, setNotificationMessage] = useState(null)
-    
+    const [notificationMessage2, setNotificationMessage2] = useState(null)
+    const [amountinCart, setamountinCart] = useState("0");
     const fetchShoppingCart = async () => {
         try {
             const response = await axios.get("http://localhost:3000/shoppingCart");
@@ -15,11 +16,21 @@ export default function ShoppingCart() {
             setError("Error fetching shoppingcart: " + (err.response?.data?.error || err.message));
         }
     };
+    const getAmountinCart = async () =>{
+        try {
+            const response = await axios.get("http://localhost:3000/shoppingCart/Count");
+            setamountinCart(response.data.count);
+        } catch (err) {
+            setError("Error fetching shoppingcart amount: " + (err.response?.data?.error || err.message));
+        }
+    }
     const deleteFromShoppingCart = async (id, name) => {
         try {
             await axios.delete("http://localhost:3000/shoppingCart/" + id);
             const response = await axios.get("http://localhost:3000/shoppingCart");
             setShoppingCart(response.data);
+            const response2 = await axios.get("http://localhost:3000/shoppingCart/Count");
+            setamountinCart(response.data.count);
             setNotificationMessage(
                 `Tuote ${name} poistettiin ostoskorista`
             )
@@ -32,12 +43,18 @@ export default function ShoppingCart() {
         }
     }
     const buyStuff =  () =>{
-        axios.delete("http://localhost:3000/shoppingCart");
-        window.location.href = "/checkout"
+        if(amountinCart > 0){
+            axios.delete("http://localhost:3000/shoppingCart");
+            window.location.href = "/checkout"
+        } else{
+            setNotificationMessage2("Virhe tuotteiden maksussa: yhtään tuotetta ei ole ostoskorissa!")
+        }
+        
 
     }
     useEffect(() => {
         fetchShoppingCart();
+        getAmountinCart();
     }, []);
 
     return (
@@ -58,6 +75,7 @@ export default function ShoppingCart() {
             {notificationMessage && <p style={{ color: "red" }}>{notificationMessage}</p>}
             
             <button className="checkout" onClick={buyStuff}>Maksa tuotteet</button>
+            {notificationMessage2 && <p style={{ color: "red" }}>{notificationMessage2}</p>}
         </div>
     );
 }
